feat: update version to 0.9.92 and implement project removal and listing functionality in separate modules
