+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nyx"
-version = "0.9.9"
+version = "0.9.92"
 edition = "2021"
 authors = ["Elouan DA COSTA PEIXOTO <elouandacostapeixoto@gmail.com>"]
 description = "A CLI tool for managing your system."
