+++ src/projects/list/mod.rs
@@ -1,0 +1,92 @@
+use crate::projects::{self};
+use crate::utils;
+use inquire::{error::InquireError, Select};
+use std::{env, fs};
+use tabled::settings::Style;
+use tabled::Table;
+
+pub fn list_projects() {
+    println!("Listing all projects...");
+    let args: Vec<String> = env::args().collect();
+    let projects = utils::get_app_vec();
+    let projects_short = utils::get_app_vec_short();
+    let mut builder = Table::builder(&projects).index().name(None);
+    if let Some(arg) = args.iter().last() {
+        match arg.as_str().trim() {
+            "-s" => {
+                builder = Table::builder(&projects_short).index().name(None);
+            }
+            "--short" => {
+                builder = Table::builder(&projects_short).index().name(None);
+            }
+            _ => {}
+        }
+    }
+    let mut table = builder.build();
+    table.with(Style::modern());
+    println!("{}", table);
+}
+
+pub fn add_existing_project_to_list() {
+    inquire::set_global_render_config(utils::get_render_config());
+    let options = utils::get_tech_option();
+    let ans: std::result::Result<String, InquireError> =
+        Select::new("Which tech your project is using ?", options).prompt();
+    match ans {
+        Ok(choice) => add_project_to_list(&choice),
+        Err(_) => println!("There was an error, please try again"),
+    }
+}
+
+pub fn add_project_to_list(tech: &String) {
+    let app_data_path = utils::get_app_data();
+    let mut projects = utils::get_app_vec();
+    let current_dir = utils::get_current_path();
+    let app_name = current_dir.split("/").last().unwrap();
+    let app_id = &app_name[..3];
+    let mut repository_user_input: String;
+    repository_user_input = utils::prompt_message(
+        "Enter the url of the github's repository of the project: ".to_string(),
+        "Failed to get the user input".to_string(),
+    );
+
+    if repository_user_input == "".to_string() {
+        repository_user_input = "No repository specified".to_string()
+    }
+
+    let mut github_project: String;
+
+    github_project = utils::prompt_message(
+        "Enter the url of the github project: ".to_string(),
+        "Error getting the user input".to_string(),
+    );
+
+    if github_project == "".to_string() {
+        github_project = "No github project specified".to_string()
+    }
+
+    let mut version: String;
+
+    version = utils::prompt_message(
+        "Enter the version of the project: ".to_string(),
+        "Error getting the user input".to_string(),
+    );
+
+    if version == "".to_string() {
+        version = "0.1.0".to_string();
+    }
+
+    let new_app: projects::Project = projects::Project {
+        id: (app_id.to_string().to_lowercase()),
+        name: (app_name.to_string()),
+        tech: (tech.to_string()),
+        location: (current_dir),
+        repository: repository_user_input,
+        github_project: github_project,
+        version: version,
+    };
+    projects.push(new_app.clone());
+    let updated_data = projects::Data { project: projects };
+    let save_json = serde_json::to_string(&updated_data).expect("Failed to serialize data");
+    fs::write(app_data_path, save_json).expect("Failed to write updated data");
+}
